update FMRIWidgetNew.py TODO: implement the workers in FMRIWidgetNew.py
